Исправление EEPROM: переменные Range/Unit изменены на byte
Проблема:
- currentRangeIndex и currentUnitIndex были типа int (2 байта)
- При записи в EEPROM адреса пересекались:
  * ADDR_RANGE=1: записывал 2 байта (1, 2)
  * ADDR_UNIT=2: записывал 2 байта (2, 3) — адреса 2 пересекались!
- При чтении currentRangeIndex получал мусор (байт от currentUnitIndex)
- На дисплее отображался '?' вместо значений диапазонов
- При переборе реле сбивалось (выход за границы массива)

Решение:
- Изменён тип на byte (1 байт): range 0-5, unit 0-4 помещаются
- Чтение: EEPROM.read() вместо EEPROM.get()
- Запись: EEPROM.write() вместо EEPROM.put()
- prevRangeIndex/prevUnitIndex изменены на byte (инициализация 255)

+++ pressure_controller/logic.cpp
@@ -12,8 +12,8 @@ const float RANGES[RANGES_COUNT] = {1.0f, 40.0f, 250.0f, 400.0f, 600.0f, 1000.0f
 // Глобальные переменные
 // ============================================================================
 
-int   currentRangeIndex = 2;
-int   currentUnitIndex  = 1;
+byte  currentRangeIndex = 2;
+byte  currentUnitIndex  = 1;
 float setpointLow       = 20.0f;
 float setpointHigh      = 80.0f;
 float hysteresis        = 2.0f;
